chore: renamed folder modules to features
