+++ src/features/common/interfaces/index.ts
